move ParentDoesNotExistError exception under bft
